feat: influxdata: initial tokenization primitives
This represents the first steps towards a (hopefully) fast and zero-
or near-zero allocation line-protocol tokenizer that can be used as the
basis for a higher level package API TBD.

The package name is, of course, subject to bikeshedding. I'm keen
that, regardless of the chosen name, the package name should
match the last element of its import path
(unlike the current `github.com/influxdata/line-protocol`, which is non-descriptly
named `protocol`).

As this is a low level API, there are a few convenience trade-offs that
are being made in anticipation of performance benefits:

- methods do not return errors. If there's a read error, that will be
picked up at a higher level - in general we want to consume as much
data as possible even in the face of a read error - and this reduces
call overhead too.

- we anticipate that several sections of the input will want to be kept
around at the same time (so we can return both a key and a value from
the same call without extra copies or allocations), hence the "from"
argument to `take` and `isEmpty`, allowing the higher level methods to
manage the contents of the buffer while consuming input.